Refactored fund amount/units handling in asset action form
Added toggle buttons to trigger auto computations: when amount (or units) is changed and there is a net asset value, the counterpart units (or amount) is computed when enabled.
Toggle button status is saved in application settings.

The net asset value field can be edited. The value is only used for following computations (i.e. it is not saved in asset history).

Computation result is scaled (and rounded) as necessary.
Amount and units scale and rounding method can be changed in the 'Options' form.

Upon transfer, the source amount value is copied (when changed) to the destination amount field.