Really resolve asset availability date
When computing assets, if scheme/fund has only one availability date in the past it was kept (as informational, since we properly display the asset as available).
This is not really useful and prevents simple computing of asset total by availaiblity date.
So even if this case forget the availability date.