Allow displaying actual VWAP or VWAP per asset
Actual VWAP is computed from invested amount of all assets in a given scheme+fund (independently of availabilities). It should be the one used when computing gains or losses (and invested amount upon transfer).
However a finer VWAP per asset (dependent of assets availabilities) can give us a finer view on gains and losses for a given operation (at least until the asset become available and merged with other assets from the same scheme+fund). So compute both, and have an option to display one or the other.