Code refactoring and simplification
Re-use existing functions or variables when possible in schemes/funds editing forms:
* use function which refreshes list view to initially display entries too
* use form status (entry ready to be modified or added/duplicated) when checking for pending changes before leaving form

Don't have a Cancel button when asking which associated funds/schemes to remove when deleting a scheme/fund.
It could be confusing since it would only cancel the associated funds/schemes deletion, not the scheme or fund itself: instead user has to de-select entries explicitely (or close the dialog) in order to not delete any associated resources.