Enhancements in account history table
Make sure to display events in history order. Use an index (not visible) from sorted events to enforce correct ordering.
Add an entry (without date) when a scheme or fund is renamed as it is useful to know (and understand what happens next, especially if old name is reused).
Display an error message when chart content building unexpectedly fails.
Disable sorting on event description, as it has no meaning.