Chart enhancements
Keep the Y axis on the left of the view when scrolling, so that it's
easier to determine the y values.
Update code so that drawn horizontal lines still go to the Y axis.

Don't reset the reference data when changing zoom.

Better computation of the zoom node, fixing off-by-1 rounding when
scrolling (which made the node visually move forth and back by 1 pixel).

Fixed most code warnings.