Prevent chart marker overlapping
When adding markers, do it in chronological order.
If a new marker would overlap (more than 1/3 of its width) with the
previous one, don't create it but re-use the previous one (updates
the tooltip text).
If we need to highlight a mark for which there is no distinct marker
(another one re-used due to overlapping), create a temporary marker
that only exists during the highlighting.