Reuse old markers by increasing date
Easier to handle than trying to re-use a marker with the same date.
When zooming (chart width increasing), this limits the number of
markers to add to the bare minimum: only re-using markers with the
same date often trigger more marker deletion/creation (due to the
way overlapping is prevented).