Handle chart 'clipping' mode
As an alternative to build a chart displaying the whole data range
in one big pane (with scrollbar to scroll inside the pane), allow
building a chart that only displays a portion (X range) of the
data, associated to a scrollbar allowing to change the viewed range
relatively to the whole data range.

The viewing mode can be changed on-the-fly through a control node
present next to the zooming node.
Default mode is 'clipping', as it is usually faster than legacy.

Collision detection (between markers and zooming node) code changed
to take into account the new control: the 'zoom' and 'clipping'
nodes are now part of a parent 'controls' node.

Other changes:
Collision code now uses the marker group (matches the actual SVG
icon shape) instead of pane to detect collision.
When changing view with tracking, remember the previous tracking if
any and cancel it before tracking again.