Better handle asset transfer with two different dates
As fas as NAV history is concerned, the filled value is for the
operation date.
The second date is when the operation is effective (truly visible)
at least on the destination fund, and possibly the source too.

For simplicity, in the scheme history we apply the operation on
the event date for both source and destination. Thus the second
date is mostly for information purposes there.

Fix destination fund NAV history updating:
 - ensure we insert/update NAV for the event date: this is
   necessary for proper NAV history, and prevents any errors when
   showing account history while destination fund had no NAV
   history up to the event date
 - if there was no NAV history up to the second date, also insert
   NAV for this second date: makes the fund NAV history better
   looking